+++ src/collective/cover/tiles/basic.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from collective.cover import _
+from collective.cover import utils
 from collective.cover.interfaces import ISearchableText
 from collective.cover.logger import logger
 from collective.cover.tiles.base import IPersistentCoverTile
@@ -26,6 +27,14 @@ class IBasicTile(IPersistentCoverTile):
 
     description = schema.Text(
         title=_(u'Description'),
+        required=False,
+    )
+
+    form.omitted(IDefaultConfigureForm, 'remote_url')
+    remote_url = schema.URI(
+        title=_(u'label_remote_url', default=u'URL'),
+        description=_(
+            u'help_remote_url', default=u'Use absolute links only.'),
         required=False,
     )
 
@@ -108,10 +117,8 @@ class BasicTile(PersistentCoverTile):
         return not [i for i in self.data.values() if i]
 
     def getURL(self):
-        """ Return the URL of the original object.
-        """
-        if self.brain:
-            return self.brain.getURL()
+        """Return the remote URL field."""
+        return self.data.get('remote_url') or u''  # deal with None values
 
     def Subject(self):
         """ Return the categories of the original object (AKA keywords, tags
@@ -138,6 +145,7 @@ class BasicTile(PersistentCoverTile):
         data = {
             'title': title,
             'description': description,
+            'remote_url': utils.get_absolute_url(obj),
             'uuid': IUUID(obj),
             'date': True,
             'subjects': True,